telegram_django/requirements.txt
Django==3.1.7
djangorestframework==3.12.2

ipdb==0.13.6
ipython==7.21.0

gunicorn==20.0.4
dj-database-url==0.5.0

python-dotenv==0.15.0